requirements.txt
google-api-python-client~=1.12.5
google_auth_oauthlib
requests~=2.24.0
appdirs~=1.4.4
python-dateutil~=2.8.1